Add enabled toggle
